v4.0

+++ option/src/app/java_helidon/rest/src/main/java/me/opc/mp/database/Dept.java
@@ -33,7 +33,7 @@ public class Dept {
         this.loc = loc;
     }
 
-    public int getDeptno() {
+    public long getDeptno() {
         return deptno;
     }
