+++ option/src/app/java_helidon4/rest/src/main/java/helidon/Dept.java
@@ -35,7 +35,7 @@ public class Dept {
         this.loc = loc;
     }
 
-    public int getDeptno() {
+    public long getDeptno() {
         return deptno;
     }
